Title: 사칙연산 함수, Time: 124ms, Memory: 24MB, Status: Passed - CodeTree

+++ 241011/사칙연산 함수/quadratic-operations-function.py
@@ -5,7 +5,7 @@ def Minus(a, c):
 def Multi(a, c):
     return a * c
 def Div(a, c):
-    return a / c
+    return int(a / c)
 
 a, o, c = input().split()
 if o == '+':
